Update flags interface for long long flags.
That was painful.

+++ examples/hmctest/hmctest5.cpp
@@ -184,7 +184,7 @@ int main( int argc, const char* argv[] )
 
     BeagleInstanceDetails instDetails;
 
-    long preferenceFlags = BEAGLE_FLAG_SCALERS_RAW;
+    long long preferenceFlags = BEAGLE_FLAG_SCALERS_RAW;
 
     if (useGpu) {
         preferenceFlags |= BEAGLE_FLAG_PROCESSOR_GPU;
@@ -198,7 +198,7 @@ int main( int argc, const char* argv[] )
         preferenceFlags |= BEAGLE_FLAG_PRECISION_DOUBLE;
     }
 
-    long requirementFlags = BEAGLE_FLAG_EIGEN_REAL;
+    long long requirementFlags = BEAGLE_FLAG_EIGEN_REAL;
     if (autoTranspose) {
         requirementFlags |= BEAGLE_FLAG_PREORDER_TRANSPOSE_AUTO;
     } else {
