+++ libhmsbeagle/GPU/BeagleGPUImpl.h
@@ -239,8 +239,8 @@ public:
                        int scaleBufferCount,
                        int resourceNumber,
                        int pluginResourceNumber,
-                       long preferenceFlags,
-                       long requirementFlags);
+                       long long preferenceFlags,
+                       long long requirementFlags);
     
     int getInstanceDetails(BeagleInstanceDetails* retunInfo);
 
@@ -528,8 +528,8 @@ public:
                                    int scaleBufferCount,
                                    int resourceNumber,
                                    int pluginResourceNumber,
-                                   long preferenceFlags,
-                                   long requirementFlags,
+                                   long long preferenceFlags,
+                                   long long requirementFlags,
                                    int* errorCode);
 
     virtual const char* getName();
